lint, refactor CavcSelectionBasis into separate component, modify serializers to return new object relations
