+++ app/models/cavc_dispositions_to_reason.rb
@@ -5,9 +5,8 @@ class CavcDispositionsToReason < CaseflowRecord
 
   belongs_to :cavc_dashboard_disposition
   belongs_to :cavc_decision_reason
-  has_many :cavc_reasons_to_bases
+  has_many :cavc_reasons_to_bases, dependent: :destroy
   has_many :cavc_selection_bases, through: :cavc_reasons_to_bases
 
   validates :cavc_dashboard_disposition, presence: true
-
 end
