+++ app/models/serializers/work_queue/cavc_dispositions_to_reason_serializer.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 class WorkQueue::CavcDispositionsToReasonSerializer
   include FastJsonapi::ObjectSerializer
 
   attribute :id
-  attribute :cavc_dashboard_disposition
-  attribute :cavc_decision_reason
-  attribute :cavc_selection_base
+  attribute :cavc_dashboard_disposition_id
+  attribute :cavc_decision_reason_id
+  attribute :cavc_reasons_to_bases
 end
